feature: First pass at Container Queries
`react-container-query` used to emulate 'media' queries based on container widths. Needs more thought on how to handle layouts for multiple devices. Min / Max widths for sections probs a good idea.

Added in an 'Edit' button to toggle editing the layout.

Added mostly static chakra elements to each of the sections to prototype the above Container Queries. I know it is not pretty, but TypeScript and I wanted to prototype it without stalling. I *will* split them out into components but would be good to perhaps pair up with you for this when you're available @vidvidvid.  🙏