adding IDX basic profile & removing erroring options in tsc config 🧣
After adding `@datamodels/identity-profile-basic` to `packages/web/`, I get an error running `yarn generate` that `tsBuildInfoFile` and `incremental` are not valid options.

I tried upgrading to `typescript` v4.4.4 and adding it to `resolutions` in `package.json`, but neither fixed it.

It only occurs when running `yarn generate` in `packages/web/` and not in the root.

Because there's a pretty severe time crunch on this, rather than taking more time, I simply removed the configuration options.

I don't know the precise repercussions, but expect the effects to solely on optimizations.