Revert previous change and put the max width on separate components
The page wrapper on players.tsx main content should not have a max-width because some of its elements need to go full-width e.g. the wrapper of the desktop filters

Max-width of 7xl applied to PlayerList.tsx, various rows made by components in PlayerFilter.tsx. The max-width used previously was 79rem, changed it to a Chakra var